SNOW-1561613: Move failing OldDriver tests with GCP to LatestIT
